refactor: 중복 코드를 부모 클래스로 이동
