+++ jongeuni/src/jongeuni/chapther10/Phone.java
@@ -7,32 +7,19 @@ import jongeuni.chapther02.Money;
 import lombok.Getter;
 
 @Getter
-public class Phone {
+public class Phone extends AbstractPhone{
 
     private Money amount;
     private Duration seconds;
-    private List<Call> calls = new ArrayList<Call>();
 
     public Phone(Money amount, Duration seconds) {
         this.amount = amount;
         this.seconds = seconds;
     }
 
-    public void call(Call call) {
-        calls.add(call);
-    }
 
-    public Money calculateFee() {
-        Money result = Money.ZERO;
-
-        for(Call call : calls) {
-            result = result.plus(calculateCallFee(call));
-        }
-
-        return result;
-    }
-
-    private Money calculateCallFee(Call call) {
+    @Override
+    protected Money calculateCallFee(Call call) {
         return amount.times(call.getDuration().getSeconds() / seconds.getSeconds());
     }
 }
